fix(prosecheck): honor terminology surfaces
